chore: tests and ci (#4)
